+++ samples/mmap-test.cpp
@@ -24,7 +24,7 @@ try {
     }
     // Map it to memory
     const Exiv2::byte* pData = file.mmap();
-    long size = file.size();
+    long size = (long)file.size();
     DataBuf buf(size);
     // Read from the memory mapped region
     memcpy(buf.pData_, pData, buf.size_);
