+++ src/tiffimage.cpp
@@ -193,7 +193,7 @@ namespace Exiv2 {
                                           iptcData_,
                                           xmpData_,
                                           io_->mmap(),
-                                          io_->size());
+                                          (uint32_t) io_->size());
         setByteOrder(bo);
 
         // read profile from the metadata
@@ -219,7 +219,7 @@ namespace Exiv2 {
             // Ensure that this is the correct image type
             if (isTiffType(*io_, false)) {
                 pData = io_->mmap(true);
-                size = io_->size();
+                size = (long) io_->size();
                 TiffHeader tiffHeader;
                 if (0 == tiffHeader.read(pData, 8)) {
                     bo = tiffHeader.byteOrder();
